+++ SherlDog/Reusable/Component/ButtonManager.swift
@@ -38,15 +38,4 @@ class ButtonManager: UIButton {
         }
     }
     
-    private func setBackgroundColor(_ color: UIColor, for state: UIButton.State) {
-        let renderer = UIGraphicsImageRenderer(size: .init(width: 1, height: 1))
-        
-        let image = renderer.image { context in
-            color.setFill()
-            context.fill(.init(origin: .zero, size: .init(width: 1, height: 1)))
-        }
-        
-        setBackgroundImage(image, for: state)
-    }
-    
 }
